상품정보 덥데이트 숮어

+++ src/main/java/com/example/ProductController.java
@@ -34,15 +34,17 @@ public class ProductController {
         return "list";
     }
 
+    public boolean checkLogin(HttpServletRequest request, HttpServletResponse response) throws Exception {
+        return loginInterceptor.preHandle(request, response, null);
+    }
+
     @GetMapping("/{itemId}")
     public String getProductDetails(@PathVariable Integer itemId, Model model, HttpServletRequest request, HttpServletResponse response) {
         try {
-            if (!loginInterceptor.preHandle(request, response, null)) {
-                // 세션이 유효하지 않은 경우 로그인 페이지로 리다이렉트
+            if (!checkLogin(request, response)) {
                 return "redirect:/login/login";
             }
 
-            // 세션이 유효한 경우에만 실행되는 코드
             ProductVO product = productDao.getProduct(itemId);
             model.addAttribute("product", product);
             return "view";
@@ -55,10 +57,12 @@ public class ProductController {
     public String addProducts(HttpServletRequest request, HttpServletResponse response) {
         try {
             if (!loginInterceptor.preHandle(request, response, null)) {
+                System.out.println("로그인 안됨");
                 // 세션이 유효하지 않은 경우 로그인 페이지로 리다이렉트
                 return "redirect:/login/login";
             }
 
+            System.out.println("로그인 됨");
             // 세션이 유효한 경우에만 실행되는 코드
             return "add";
         } catch (Exception e) {
@@ -79,14 +83,48 @@ public class ProductController {
         return "redirect:/products";
     }
 
-    @GetMapping("/delete/{id}")
-    public String deletePost(@PathVariable int id) {
-        if (productDao.deleteProduct(id) != 0) {
-            System.out.println("success");
-        } else {
-            System.out.println("fail");
-        }
+    @GetMapping("/edit/{id}")
+    public String editPost(@PathVariable int id, Model model, HttpServletRequest request, HttpServletResponse response) {
+        try {
+            if (!loginInterceptor.preHandle(request, response, null)) {
+                // 세션이 유효하지 않은 경우 로그인 페이지로 리다이렉트
+                return "redirect:/login/login";
+            }
 
-        return "redirect:../../products";
+            // 세션이 유효한 경우에만 실행되는 코드
+            ProductVO product = productDao.getProduct(id);
+            model.addAttribute("product", product);
+            return "edit";
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
     }
+
+    @PostMapping("/editOk")
+    public String updateProducts(ProductVO productVO) {
+        productDao.updateProduct(productVO);
+        return "redirect:/products";
+    }
+
+    @GetMapping("/delete/{id}")
+    public String deletePost(@PathVariable int id, HttpServletRequest request, HttpServletResponse response) {
+        try {
+            if (!loginInterceptor.preHandle(request, response, null)) {
+                // 세션이 유효하지 않은 경우 로그인 페이지로 리다이렉트
+                return "redirect:/login/login";
+            }
+
+            // 세션이 유효한 경우에만 실행되는 코드
+            if (productDao.deleteProduct(id) != 0) {
+                System.out.println("success");
+            } else {
+                System.out.println("fail");
+            }
+
+            return "redirect:../../products";
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
 }
